+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v1.0.0
 - Seal `BitDecodeExt`, `BitEncodeExt`, `BitCodec` traits
 - Add `#[codec(other)]` attribute
+- Replace `#[codec(default)]` with `#[codec(skip_encode)]`, `#[codec(skip_decode)]`, `#[codec(skip)]`
+- Modify `Discriminable::discriminant` to return `Option`
 # v0.11.1
 - Fix potential memory leak in `BitDecode` implementation for `[T; N]`
 # v0.11.0